--- Assets/Scripts/TileManagement/OldTileGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 37f6f7f1aaa8e60469e5a237b3813d50
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
